refactor: rename external tool URL methods and interfaces for clarity
