+++ src/externalTools/domain/repositories/ExternalToolsRepository.ts
@@ -1,16 +1,16 @@
 import { DatasetExternalToolResolved } from '../models/DatasetExternalToolResolved'
 import { ExternalTool } from '../models/ExternalTool'
 import { FileExternalToolResolved } from '../models/FileExternalToolResolved'
-import { GetExternalToolDTO } from '../useCases/DTOs/GetExternalToolUrlDTO'
+import { GetExternalToolDTO } from '../useCases/DTOs/GetExternalToolDTO'
 
 export interface ExternalToolsRepository {
   getExternalTools(): Promise<ExternalTool[]>
-  getDatasetExternalToolUrl(
+  getDatasetExternalToolResolved(
     datasetId: number | string,
     toolId: number,
     getExternalToolDTO: GetExternalToolDTO
   ): Promise<DatasetExternalToolResolved>
-  getFileExternalToolUrl(
+  getFileExternalToolResolved(
     fileId: number | string,
     toolId: number,
     getExternalToolDTO: GetExternalToolDTO
