+++ src/externalTools/domain/useCases/DTOs/GetExternalToolDTO.ts
