+++ src/externalTools/domain/useCases/GetDatasetExternalToolResolved.ts
@@ -1,12 +1,16 @@
 import { DatasetExternalToolResolved } from '../models/DatasetExternalToolResolved'
 import { ExternalToolsRepository } from '../repositories/ExternalToolsRepository'
-import { GetExternalToolDTO } from './DTOs/GetExternalToolUrlDTO'
+import { GetExternalToolDTO } from './DTOs/GetExternalToolDTO'
 
-export function getDatasetExternalToolUrl(
+export function getDatasetExternalToolResolved(
   externalToolsRepository: ExternalToolsRepository,
   datasetId: number | string,
   toolId: number,
   getExternalToolDTO: GetExternalToolDTO
 ): Promise<DatasetExternalToolResolved> {
-  return externalToolsRepository.getDatasetExternalToolUrl(datasetId, toolId, getExternalToolDTO)
+  return externalToolsRepository.getDatasetExternalToolResolved(
+    datasetId,
+    toolId,
+    getExternalToolDTO
+  )
 }
