+++ src/externalTools/domain/useCases/GetFileExternalToolResolved.ts
@@ -1,12 +1,12 @@
 import { FileExternalToolResolved } from '../models/FileExternalToolResolved'
 import { ExternalToolsRepository } from '../repositories/ExternalToolsRepository'
-import { GetExternalToolDTO } from './DTOs/GetExternalToolUrlDTO'
+import { GetExternalToolDTO } from './DTOs/GetExternalToolDTO'
 
-export function getFileExternalToolUrl(
+export function getFileExternalToolResolved(
   externalToolsRepository: ExternalToolsRepository,
   fileId: number | string,
   toolId: number,
   getExternalToolDTO: GetExternalToolDTO
 ): Promise<FileExternalToolResolved> {
-  return externalToolsRepository.getFileExternalToolUrl(fileId, toolId, getExternalToolDTO)
+  return externalToolsRepository.getFileExternalToolResolved(fileId, toolId, getExternalToolDTO)
 }
