+++ src/externalTools/infrastructure/repositories/ExternalToolsJSDataverseRepository.ts
@@ -1,36 +1,36 @@
 import {
-  getDatasetExternalToolUrl,
+  getDatasetExternalToolResolved,
   getExternalTools,
-  getFileExternalToolUrl
+  getFileExternalToolResolved
 } from '@iqss/dataverse-client-javascript'
 import { DatasetExternalToolResolved } from '@/externalTools/domain/models/DatasetExternalToolResolved'
 import { ExternalTool } from '@/externalTools/domain/models/ExternalTool'
 import { FileExternalToolResolved } from '@/externalTools/domain/models/FileExternalToolResolved'
 import { ExternalToolsRepository } from '@/externalTools/domain/repositories/ExternalToolsRepository'
-import { GetExternalToolDTO } from '@/externalTools/domain/useCases/DTOs/GetExternalToolUrlDTO'
+import { GetExternalToolDTO } from '@/externalTools/domain/useCases/DTOs/GetExternalToolDTO'
 
 export class ExternalToolsJSDataverseRepository implements ExternalToolsRepository {
   getExternalTools(): Promise<ExternalTool[]> {
     return getExternalTools.execute().then((jsExternalTools) => jsExternalTools)
   }
 
-  getDatasetExternalToolUrl(
+  getDatasetExternalToolResolved(
     datasetId: number | string,
     toolId: number,
     getExternalToolDTO: GetExternalToolDTO
   ): Promise<DatasetExternalToolResolved> {
-    return getDatasetExternalToolUrl
+    return getDatasetExternalToolResolved
       .execute(datasetId, toolId, getExternalToolDTO)
-      .then((jsDatasetExternalToolUrl) => jsDatasetExternalToolUrl)
+      .then((jsDatasetExternalToolResolved) => jsDatasetExternalToolResolved)
   }
 
-  getFileExternalToolUrl(
+  getFileExternalToolResolved(
     fileId: number | string,
     toolId: number,
     getExternalToolDTO: GetExternalToolDTO
   ): Promise<FileExternalToolResolved> {
-    return getFileExternalToolUrl
+    return getFileExternalToolResolved
       .execute(fileId, toolId, getExternalToolDTO)
-      .then((jsFileExternalToolUrl) => jsFileExternalToolUrl)
+      .then((jsFileExternalToolResolved) => jsFileExternalToolResolved)
   }
 }
